requirements.txt
paho.mqtt
pyserial
crcmod
pycryptodomex
pycryptodome